Organised Java files to separate directories
