+++ README.md
@@ -1,2 +1,2 @@
-# Safepassage Android
 ![Android CI](https://github.com/VICS-CORE/safepassage_android/workflows/Android%20CI/badge.svg)
+# Safepassage Android
