+++ app/src/PassIssuer/java/org/covid19india/android/safepassageindia/passissuer/activity/MenuActivity.java
@@ -1,9 +1,12 @@
-package org.covid19india.android.safepassageindia.passissuer;
+package org.covid19india.android.safepassageindia.passissuer.activity;
 
 import android.os.Bundle;
+import android.util.Log;
 
+import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.material.tabs.TabLayout;
 import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.GetTokenResult;
 
 import org.covid19india.android.safepassageindia.R;
 import org.covid19india.android.safepassageindia.passissuer.adapter.SectionsPagerAdapter;
@@ -19,6 +22,7 @@ import androidx.fragment.app.FragmentPagerAdapter;
 import androidx.viewpager.widget.ViewPager;
 
 public class MenuActivity extends AppCompatActivity {
+    private static final String TAG = "MenuActivity";
     ViewPager viewPager;
     TabLayout tabLayout;
     SectionsPagerAdapter pagerAdapter;
@@ -29,6 +33,12 @@ public class MenuActivity extends AppCompatActivity {
         setContentView(R.layout.activity_menu);
         init();
         setViewPager();
+        FirebaseAuth.getInstance().getCurrentUser().getIdToken(false).addOnSuccessListener(new OnSuccessListener<GetTokenResult>() {
+            @Override
+            public void onSuccess(GetTokenResult getTokenResult) {
+                Log.d(TAG, "TokenId = " + getTokenResult.getToken());
+            }
+        });
         /*textView = findViewById(R.id.welcome_text);
         textView.setText(FirebaseAuth.getInstance().getCurrentUser().getPhoneNumber());
         button = findViewById(R.id.sign_out);
@@ -45,9 +55,9 @@ public class MenuActivity extends AppCompatActivity {
     private void setViewPager() {
         List<Fragment> fragmentList = new ArrayList<>();
         List<String> titleList = new ArrayList<>();
-        fragmentList.add(PassesFragment.newInstance("Similar to intent.putExtra method","Similar to intent.putExtra method"));
+        fragmentList.add(PassesFragment.newInstance("Similar to intent.putExtra method", "Similar to intent.putExtra method"));
         titleList.add("Passes");
-        fragmentList.add(TeamsFragment.newInstance("Similar to intent.putExtra method","Similar to intent.putExtra method"));
+        fragmentList.add(TeamsFragment.newInstance("Similar to intent.putExtra method", "Similar to intent.putExtra method"));
         titleList.add("Teams");
 
         pagerAdapter.setFragments(fragmentList, titleList);
