+++ app/src/PassIssuer/res/layout/activity_menu.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:background="@color/colorBackground"
     android:fitsSystemWindows="true"
-    tools:context="org.covid19india.android.safepassageindia.passissuer.MenuActivity">
+    tools:context="org.covid19india.android.safepassageindia.passissuer.activity.MenuActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/appBarLayout"
