+++ app/src/PassIssuer/res/layout/item_list.xml
@@ -12,18 +12,18 @@
         android:id="@+id/textView_name"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_toRightOf="@id/imageView_poster"
         android:layout_margin="20dp"
+        android:layout_toEndOf="@id/imageView_poster"
         android:textSize="20sp"
-        android:textStyle="bold"/>
+        android:textStyle="bold" />
 
     <TextView
         android:id="@+id/textView_role"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_toRightOf="@id/imageView_poster"
         android:layout_below="@id/textView_name"
         android:layout_margin="20dp"
-        android:textSize="20sp"/>
+        android:layout_toEndOf="@id/imageView_poster"
+        android:textSize="20sp" />
 
 </RelativeLayout>
