+++ app/src/PassScanner/java/org/covid19india/android/safepassageindia/passscanner/activity/MenuActivity.java
@@ -1,20 +1,24 @@
-package org.covid19india.android.safepassageindia.passscanner;
+package org.covid19india.android.safepassageindia.passscanner.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
+import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.GetTokenResult;
 
 import org.covid19india.android.safepassageindia.BroadcastReceiverService;
-import org.covid19india.android.safepassageindia.LoginActivity;
 import org.covid19india.android.safepassageindia.R;
+import org.covid19india.android.safepassageindia.activity.LoginActivity;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 public class MenuActivity extends AppCompatActivity {
+    private static final String TAG = "MenuActivity";
     TextView textView;
     Button scanButton, signOutButton;
 
@@ -23,6 +27,12 @@ public class MenuActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_menu);
         init();
+        FirebaseAuth.getInstance().getCurrentUser().getIdToken(false).addOnSuccessListener(new OnSuccessListener<GetTokenResult>() {
+            @Override
+            public void onSuccess(GetTokenResult getTokenResult) {
+                Log.d(TAG, "TokenId = " + getTokenResult.getToken());
+            }
+        });
         String welcomeMessage = "Welcome " + FirebaseAuth.getInstance().getCurrentUser().getPhoneNumber();
         textView.setText(welcomeMessage);
         scanButton.setOnClickListener(new View.OnClickListener() {
