+++ app/src/PassScanner/java/org/covid19india/android/safepassageindia/passscanner/activity/ScannerActivity.java
@@ -1,4 +1,4 @@
-package org.covid19india.android.safepassageindia.passscanner;
+package org.covid19india.android.safepassageindia.passscanner.activity;
 
 import android.content.Intent;
 import android.content.pm.PackageManager;
