+++ app/src/PassScanner/java/org/covid19india/android/safepassageindia/passscanner/adapter/PassAdapter.java
@@ -1,4 +1,4 @@
-package org.covid19india.android.safepassageindia.passscanner;
+package org.covid19india.android.safepassageindia.passscanner.adapter;
 
 import android.content.Context;
 import android.content.Intent;
@@ -8,8 +8,8 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import org.covid19india.android.safepassageindia.PassViewActivity;
 import org.covid19india.android.safepassageindia.R;
+import org.covid19india.android.safepassageindia.activity.PassViewActivity;
 import org.covid19india.android.safepassageindia.model.UserPassList;
 
 import androidx.annotation.NonNull;
