+++ app/src/PassUser/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application>
-        <activity android:name="org.covid19india.android.safepassageindia.passuser.MenuActivity"></activity>
+        <activity android:name="org.covid19india.android.safepassageindia.passuser.activity.MenuActivity"></activity>
     </application>
 </manifest>
