+++ app/src/main/AndroidManifest.xml
@@ -12,14 +12,14 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".PassViewActivity"></activity>
+        <activity android:name=".activity.PassViewActivity"></activity>
 
         <service
             android:name=".BroadcastReceiverService"
             android:enabled="true"
             android:exported="true" />
 
-        <activity android:name=".LoginActivity">
+        <activity android:name=".activity.LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
