+++ app/src/main/java/org/covid19india/android/safepassageindia/model/UserList.java
@@ -2,13 +2,11 @@ package org.covid19india.android.safepassageindia.model;
 
 import com.google.gson.annotations.SerializedName;
 
-import org.covid19india.android.safepassageindia.model.User;
-
 import java.util.List;
 
 public class UserList {
     @SerializedName("user")
-    List<User> users;
+    private List<User> users;
 
     public UserList(List<User> users) {
         this.users = users;
